requirements.txt
dash
plotly
pandas
biopython
requests
dash-bootstrap-components
